+++ lib/multi_image_picker.dart
@@ -1,2 +1,3 @@
-export 'picker.dart';
+export 'asset.dart';
 export 'cupertino_options.dart';
+export 'picker.dart';